transfer col2im to image space
