+++ python/tvm/relay/op/image/image.py
@@ -372,6 +372,53 @@ def crop_and_resize(
     )
 
 
+def col2im(data, image_shape, block_shape, dilation, pads, strides):
+    """Rearrange column blocks back into a multidimensional image.
+
+    Parameters
+    ----------
+    data : relay.Expr
+        The input data tensor to be rearranged from column blocks back into an image.
+        This is a 3-D tensor containing [N, C * n-ary-product(block_shape), L],
+        where N is batch dimension, C is image channel dimension and L is number of blocks.
+
+    image_shape : relay.Expr
+        The shape of the spatial dimensions of the image after rearranging the column blocks.
+        This is a 1-D tensor with size of at least 2, containing the value [H_img, W_img]
+        for a 2-D image or [dim_i1, dim_i2, ..., dim_iN] for a N-D image.
+
+    block_shape : relay.Expr
+        The shape of the block to apply on the input. This is a 1-D tensor of size of
+        at least 2, containing the value [H_block, W_block] for a 2-D image or
+        [dim_b1, dim_b2, ..., dim_bN] for a N-D block.
+        This is the block-shape before dilation is applied to it.
+
+    dilation : List[int]
+        1-D tensor with dilation value along each spatial axis of the image.
+        If not present, the dilation defaults to 1 along each spatial axis of the image.
+
+    pads : List[int]
+        1-D tensor with padding value for the beginning and ending along each spatial axis,
+        it can take any value greater than or equal to 0. The value represent the number of
+        pixels added to the beginning and end part of the corresponding axis.
+        `pads` format should be as follow [x1_begin, x2_begin...x1_end, x2_end,...],
+        where xi_begin is the number of pixels added at the beginning of axis `i` and
+        xi_end is the number of pixels added at the end of axis `i`.
+        If not present, the padding defaults to 0 along start and end of each spatial axis.
+
+    strides : List[int]
+        1-D tensor with stride value along each spatial axis.
+        If not present, the stride defaults to 1 along each spatial axis.
+
+    Returns
+    -------
+    result : relay.Expr
+        Output tensor produced by rearranging blocks into an image.
+    """
+
+    return _make.col2im(data, image_shape, block_shape, dilation, pads, strides)
+
+
 def dilation2d(
     data,
     weight,
